Setup external testlisteners.
We now have external testlisteners, using the standard Java Service Provider mechanism.

The service (`ITestReporter`) is located in the new `org.metaborg.spt.listener` project.
The jar of the listener project (obtained via the ant build file) should be on the classpath when executing spt.

An example implementation of the `ITestReporter` service is located in the new `org.metaborg.spt.listener.junitxml` project.
This shows how `META-INF/services` is used to register the implementation.
To get this listener to output junit xml files, just build it to get the jar and add it to the classpath when running spt.